+++ src/components/react_canvas/reactive_chart.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { Layer, Rect, Stage } from 'react-konva';
 import { isLineAnnotation, isRectAnnotation, Position } from '../../lib/series/specs';
 import { LineAnnotationStyle, RectAnnotationStyle } from '../../lib/themes/theme';
-import { AnnotationId } from '../../lib/utils/ids';
+import { AnnotationId } from '../../utils/ids';
 import { AnnotationDimensions, AnnotationLineProps, AnnotationRectProps } from '../../state/annotation_utils';
 import { ChartStore, Point } from '../../state/chart_state';
 import { BrushExtent } from '../../state/utils';
