+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './specs';
 export { Chart } from './components/chart';
 export { TooltipType, TooltipValue, TooltipValueFormatter } from './lib/utils/interactions';
-export { getAxisId, getGroupId, getSpecId, getAnnotationId } from './lib/utils/ids';
+export { getAxisId, getGroupId, getSpecId, getAnnotationId } from './utils/ids';
 export { ScaleType } from './lib/utils/scales/scales';
 export { Position, Rendering, Rotation } from './lib/series/specs';
 export * from './lib/themes/theme';
