+++ src/lib/axes/axis_utils.test.ts
@@ -2,7 +2,7 @@ import { XDomain } from '../series/domains/x_domain';
 import { YDomain } from '../series/domains/y_domain';
 import { AxisSpec, DomainRange, Position, AxisStyle } from '../series/specs';
 import { LIGHT_THEME } from '../themes/light_theme';
-import { AxisId, getAxisId, getGroupId, GroupId } from '../utils/ids';
+import { AxisId, getAxisId, getGroupId, GroupId } from '../../utils/ids';
 import { ScaleType } from '../utils/scales/scales';
 import {
   AxisTick,
