+++ src/lib/series/nonstacked_series_utils.test.ts
@@ -1,4 +1,4 @@
-import { getSpecId } from '../utils/ids';
+import { getSpecId } from '../../utils/ids';
 import { formatNonStackedDataSeriesValues } from './nonstacked_series_utils';
 import { RawDataSeries } from './series';
