+++ src/lib/series/rendering.bars.test.ts
@@ -1,6 +1,6 @@
 import { computeSeriesDomains } from '../../state/utils';
 import { identity } from '../../utils/commons';
-import { getGroupId, getSpecId, SpecId } from '../utils/ids';
+import { getGroupId, getSpecId, SpecId } from '../../utils/ids';
 import { ScaleType } from '../utils/scales/scales';
 import { renderBars } from './rendering';
 import { computeXScale, computeYScales } from './scales';
