+++ src/lib/series/rendering.lines.test.ts
@@ -1,7 +1,7 @@
 /* eslint @typescript-eslint/no-object-literal-type-assertion: off */
 
 import { computeSeriesDomains } from '../../state/utils';
-import { getGroupId, getSpecId, SpecId } from '../utils/ids';
+import { getGroupId, getSpecId, SpecId } from '../../utils/ids';
 import { ScaleType } from '../utils/scales/scales';
 import { CurveType } from './curves';
 import { IndexedGeometry, LineGeometry, PointGeometry, renderLine } from './rendering';
