+++ src/lib/series/series.test.ts
@@ -1,5 +1,5 @@
 import { ColorConfig } from '../themes/theme';
-import { getGroupId, getSpecId, SpecId } from '../utils/ids';
+import { getGroupId, getSpecId, SpecId } from '../../utils/ids';
 import { ScaleType } from '../utils/scales/scales';
 import {
   DataSeriesColorsValues,
