+++ src/lib/series/specs.ts
@@ -8,7 +8,7 @@ import {
 } from '../themes/theme';
 import { Accessor } from '../utils/accessor';
 import { Omit, RecursivePartial } from '../../utils/commons';
-import { AnnotationId, AxisId, GroupId, SpecId } from '../utils/ids';
+import { AnnotationId, AxisId, GroupId, SpecId } from '../../utils/ids';
 import { ScaleContinuousType, ScaleType } from '../utils/scales/scales';
 import { CurveType } from './curves';
 import { DataSeriesColorsValues } from './series';
