+++ src/lib/series/tooltip.test.ts
@@ -1,4 +1,4 @@
-import { getAxisId, getGroupId, getSpecId } from '../utils/ids';
+import { getAxisId, getGroupId, getSpecId } from '../../utils/ids';
 import { ScaleType } from '../utils/scales/scales';
 import { BarGeometry } from './rendering';
 import { AxisSpec, BarSeriesSpec, Position } from './specs';
