+++ src/specs/area_series.tsx
@@ -1,7 +1,7 @@
 import { inject } from 'mobx-react';
 import { PureComponent } from 'react';
 import { AreaSeriesSpec, HistogramModeAlignments } from '../lib/series/specs';
-import { getGroupId } from '../lib/utils/ids';
+import { getGroupId } from '../utils/ids';
 import { ScaleType } from '../lib/utils/scales/scales';
 import { SpecProps } from './specs_parser';
