+++ src/state/annotation_marker.test.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { AnnotationDomainType, AnnotationDomainTypes, AnnotationSpec, Position, Rotation } from '../lib/series/specs';
 import { DEFAULT_ANNOTATION_LINE_STYLE } from '../lib/themes/theme';
 import { Dimensions } from '../lib/utils/dimensions';
-import { getAnnotationId, getGroupId, GroupId } from '../lib/utils/ids';
+import { getAnnotationId, getGroupId, GroupId } from '../utils/ids';
 import { ScaleContinuous } from '../lib/utils/scales/scale_continuous';
 import { Scale, ScaleType } from '../lib/utils/scales/scales';
 import {
