+++ src/state/annotation_utils.ts
@@ -18,7 +18,7 @@ import {
 } from '../lib/series/specs';
 import { LineAnnotationStyle } from '../lib/themes/theme';
 import { Dimensions } from '../lib/utils/dimensions';
-import { AnnotationId, AxisId, GroupId } from '../lib/utils/ids';
+import { AnnotationId, AxisId, GroupId } from '../utils/ids';
 import { Scale, ScaleType } from '../lib/utils/scales/scales';
 import { Point } from './chart_state';
 import { computeXScaleOffset, getAxesSpecForSpecId, isHorizontalRotation } from './utils';
