+++ src/state/chart_state.timescales.test.ts
@@ -2,7 +2,7 @@ import { DateTime } from 'luxon';
 import { LineSeriesSpec } from '../lib/series/specs';
 import { LIGHT_THEME } from '../lib/themes/light_theme';
 import { mergeWithDefaultTheme } from '../lib/themes/theme';
-import { getGroupId, getSpecId } from '../lib/utils/ids';
+import { getGroupId, getSpecId } from '../utils/ids';
 import { ScaleType } from '../lib/utils/scales/scales';
 import { ChartStore } from './chart_state';
