+++ src/state/crosshair_utils.linear_snap.test.ts
@@ -1,7 +1,7 @@
 import { computeXScale } from '../lib/series/scales';
 import { BasicSeriesSpec } from '../lib/series/specs';
 import { Dimensions } from '../lib/utils/dimensions';
-import { getGroupId, getSpecId } from '../lib/utils/ids';
+import { getGroupId, getSpecId } from '../utils/ids';
 import { ScaleType } from '../lib/utils/scales/scales';
 import { getCursorBandPosition, getSnapPosition } from './crosshair_utils';
 import { computeSeriesDomains } from './utils';
