+++ src/state/test/interactions.test.ts
@@ -2,7 +2,7 @@ import { BarGeometry } from '../../lib/series/rendering';
 import { computeXScale, computeYScales } from '../../lib/series/scales';
 import { DataSeriesColorsValues } from '../../lib/series/series';
 import { BarSeriesSpec, BasicSeriesSpec, RectAnnotationSpec } from '../../lib/series/specs';
-import { getAnnotationId, getGroupId, getSpecId } from '../../lib/utils/ids';
+import { getAnnotationId, getGroupId, getSpecId } from '../../utils/ids';
 import { TooltipType } from '../../lib/utils/interactions';
 import { ScaleContinuous } from '../../lib/utils/scales/scale_continuous';
 import { ScaleType } from '../../lib/utils/scales/scales';
